Reorganized test data
