+++ src/event/v10/serde.rs
@@ -46,12 +46,7 @@ impl crate::event::serde::EventDeserializer for EventDeserializer {
         Ok(match (data, data_base64, is_json) {
             (Some(d), None, true) => Some(Data::Json(parse_data_json!(d, E)?)),
             (Some(d), None, false) => Some(Data::String(parse_data_string!(d, E)?)),
-            (None, Some(d), true) => {
-                let data = parse_data_base64!(d, E)?;
-                Some(Data::Json(
-                    serde_json::from_slice(&data).map_err(|e| E::custom(e))?,
-                ))
-            }
+            (None, Some(d), true) => Some(Data::Json(parse_json_data_base64!(d, E)?)),
             (None, Some(d), false) => Some(Data::Binary(parse_data_base64!(d, E)?)),
             (Some(_), Some(_), _) => Err(E::custom("Cannot have both data and data_base64 field"))?,
             (None, None, _) => None,
